Move LGV Header into separate file
